refactor/dev: replace BoardManager's check methods with CheckValidator for improved clarity and maintainability
