+++ src/components/Board/Board.module.css
@@ -1,7 +1,7 @@
 .board {
     display: flex;
     flex-flow: column;
-    gap: 1rem;
+    gap: 0rem;
     max-width: 95svh;
     max-height: 95svh;
     padding: 1rem;
@@ -15,7 +15,7 @@
     display: flex;
     max-height: 5%;
     flex-flow: row nowrap;
-    gap: 1rem;
+    gap: 0rem;
     place-content: center;
 }
 
@@ -25,25 +25,20 @@
     display: flex;
     place-content: center;
     place-items: center;
-    box-shadow: 1px 1px 1rem black;
     padding: 0.33rem;
     border: 4px transparent;
 }
 
 .white {
-    background-color: white;
-    background-image: linear-gradient(331deg, #fff5c8, rgb(255, 239, 217));
-    color: black;
+    background-color: #f1dbbb;
 }
 
 .black {
-    background-color: black;
-    background-image: linear-gradient(331deg, #534633, rgb(95, 71, 50));
-    color: white;
+    background-color: #534633;
 }
 
 .highlight {
-    border: 4px dashed black;
+    filter: grayscale(15%) contrast(.75) hue-rotate(60deg);
 }
 
 .inCheck {
